pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmguardian"
version = "1.0.0"
description = "A comprehensive security toolkit for LLM applications"
authors = [{name = "dewitt4"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
dynamic = ["keywords"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.1",
    "psutil>=5.9.0",
    "python-json-logger>=2.0.7",
    "typing-extensions>=4.5.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "prometheus-client>=0.17.0",
    "statsd>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.9.1",
    "flake8>=6.1.0",
    "mypy>=1.5.1",
    "isort>=5.12.0"
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1"
]
dashboard = [
    "streamlit>=1.24.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0"
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]

[project.urls]
Homepage = "https://github.com/dewitt4/llmguardian"
Documentation = "https://llmguardian.readthedocs.io"
Repository = "https://github.com/dewitt4/llmguardian.git"
Issues = "https://github.com/dewitt4/llmguardian/issues"

[project.scripts]
llmguardian = "llmguardian.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3